Renames first file in multifile generation to use 1 of # convention
I tried to do this years ago, but it "open in nc editor" would open a log file instead of the first file.  Finally figured out how to get it to open the first file.  In order for it it work properly you need to untick "open in nc editor" which doesnt make any sense but it does work.